Extracted 'OnboardingImage' View to its own file
